Removal of the Ottoman Disaster
Postponing this to work on other projects as have hit bloc and desire not to work on ti anymore.

Wanting to move focus onto flavour for England, Holland, Pope etc. with other team. Will return work on Ottoman Stagnation when the passion is there and not got "writers block" on the disaster.